добавлены обработчики исключений в контроллерах

+++ src/main/java/carTravel/controller/CarController.java
@@ -3,7 +3,9 @@ package carTravel.controller;
 import carTravel.dto.CarDto;
 import carTravel.service.carServise.CarRepositoryService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -22,13 +24,18 @@ public class CarController {
 
     @DeleteMapping("/{id}")
     public void delete(
-            @PathVariable("id") int id) {
-        carRepositoryService.delete(id);
+            @PathVariable("id") Integer id) {
+        if (id != null) {
+            carRepositoryService.delete(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "--> НЕСУЖЕСТВУЮЩИЙ ID");
+        }
     }
 
     @GetMapping("/{id}")
     public Optional<CarDto> findById(@PathVariable int id) {
-        return carRepositoryService.get(id);
+        return Optional.ofNullable(carRepositoryService.get(id))
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "--> ПО  ДАННОМУ ID НЕ НАЙДЕНО МАШИНЫ"));
     }
 
     @GetMapping("/find-all")
