+++ src/main/java/carTravel/controller/IssueController.java
@@ -5,7 +5,9 @@ import carTravel.dto.issues.IssueSaveDto;
 import carTravel.entity.Users;
 import carTravel.service.issueServise.IssueRepositoryService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -22,19 +24,25 @@ public class IssueController {
         issueRepository.save(entity);
     }
 
-    @DeleteMapping("{id}")
+    @DeleteMapping("/{id}")
     public void delete(@PathVariable Integer id) {
-        issueRepository.deleteById(id);
+        if (id != null) {
+            issueRepository.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "--> НЕСУЖЕСТВУЮЩИЙ ID");
+        }
     }
 
     @GetMapping("/{id}")
     public Optional<IssueGetDto> findById(@PathVariable Integer id) {
-        return issueRepository.get(id);
+
+        return Optional.ofNullable(issueRepository.get(id)).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "--> ПО ДАННОМУ ID НЕ НАЙДЕНО ЗАДАЧ"));
     }
 
     @GetMapping("/get-by-executor/{id}")
     public List<IssueGetDto> findByExecutor(@PathVariable Users id) {
-        return issueRepository.getIssueByExecutor(id);
+
+        return Optional.ofNullable(issueRepository.getIssueByExecutor(id)).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "--> ПО ДАННОМУ ID НЕ НАЙДЕН ИСПОЛНИТЕЛЬ"));
     }
 
     @GetMapping("/actual-issues")
